api/chatbot-finsa/requirements.txt
fastapi==0.115.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
slowapi==0.1.9

langgraph==0.2.70
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.7.0
pydantic_core==2.23.4
SQLAlchemy==2.0.36

